Command/landmark_recognition/requirements.txt
absl-py==0.11.0
aniso8601==8.1.0
argh==0.26.2
astor==0.8.1
bokeh==2.0.2
boto==2.49.0
boto3==1.10.47
botocore==1.13.50
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
coverage==5.3.1
cryptography==3.4.6
cycler==0.10.0
dask==2.15.0
decorator==4.4.2
docker==4.4.2
docutils==0.15.2
english==2020.7.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-url-map-serializer==0.0.1
gast==0.2.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.14
grpcio==1.35.0
grpcio-tools==1.35.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.1
monotonic==1.5
mpu==0.20.0
neobolt==1.7.17
neotime==1.7.4
networkx==2.5
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==20.8
pandas==1.0.3
pansi==2020.7.3
passlib==1.7.4
pika==1.1.0
Pillow==7.1.2
prompt-toolkit==2.0.10
protobuf==3.14.0
psutil==5.7.0
psycopg2-binary==2.8.5
py2neo==2020.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pymilvus==0.2.14
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-probability==0.9.0
termcolor==1.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.46.0
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.24.3
watchdog==1.0.2
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
